# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pirate Proof - Delivery command center and smart lock API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications/push-token:
    post:
      operationId: registerPushToken
      tags: [notifications]
      summary: Register or update device push token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenInput"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Notification:
      type: object
      required: [id, title, body, type, isRead, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [delivery_update, lock_event, system, email_sync]
        isRead:
          type: boolean
        deliveryId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    PushTokenInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Expo push token (ExponentPushToken[...])
...